+++ apps/macroy_web/lib/macroy_web/controllers/orgfile_controller.ex
@@ -5,4 +5,9 @@ defmodule MacroyWeb.OrgFileController do
     orgfiles = Todo.list_org_files()
     render(conn, "index.html", orgfiles: orgfiles)
   end
+
+  def show(conn, %{"id" => id}) do
+    orgfile = Todo.get_org_file(id)
+    render(conn, "show.html", orgfile: orgfile)
+  end
 end